+++ CHANGELOG.md
@@ -19,6 +19,7 @@ for open development tasks and limitations.
 - cmd-line `scan` collection 
 - added `alembic` database migration
 - added `config` sub-command with `-db-init` and `-db-migrate`
+- `find` filter on hashtag(s)
 -